use uintptr_t in putptr

+++ pf_putptr.c
@@ -6,7 +6,7 @@
 /*   By: damateos <damateos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 22:12:58 by damateos          #+#    #+#             */
-/*   Updated: 2024/05/13 22:26:00 by damateos         ###   ########.fr       */
+/*   Updated: 2024/05/17 16:56:51 by damateos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 int	pf_putptr(void *ptr)
 {
-	unsigned long long	address;
+	uintptr_t	address;
 	int					count;
 
 	count = 0;
-	address = (unsigned long long)ptr;
+	address = (uintptr_t)ptr;
 	count += pf_putstr("0x");
 	count += pf_puthex(address, 0);
 	return (count);
